Hero basic format and fonts
