+++ src/Components/Hero/hero.module.css
@@ -1,0 +1,20 @@
+.header {
+	font-size: 4rem;
+	margin: 10rem 0;
+}
+
+.meddon {
+	font-family: "Meddon";
+}
+
+.MrsSaintDelafield {
+	font-family: "MrsSaintDelafield";
+}
+
+.center {
+	margin: 0 auto;
+}
+
+.textCenter {
+	text-align: center;
+}
