+++ src/Pages/page.module.css
@@ -3,4 +3,9 @@
 	max-width: 50rem;
 }
 
-
+.backButton {
+	position: absolute;
+	top: 0;
+	left: 0;
+	padding: 1rem;
+}
